api: add support for the MediaStream constructor
The implementation tries to be as faithful to the standard as possible,
however an extra function call is still required for cleanup:
MediaStream.release().

Now that a track can outlive the stream that originated it, there is
also MediaStreamTrack.release(). Generally it's not needed to call it
unless you have already released the stream and want to dispose of the
track alone.

Add / remove track operations on streams are only supported for local
tracks (ie, those gotten via getUserMedia).

Note: calling release() on a track still attached to a stream is
considered undefined behavior and may lead to crashes.