core: refactor datachannel creation
Data channels are now created synchronously, in order to mimic browser
behavbior.

Channel IDs are no longer the unique identifier of data channel objects
we hold on to. A UUID is used instead. This facilitated the
implementation of RTCDataChannel.id as per the spec: it may start as
null (internally -1) and it will change once negotiated.

Fixes:
https://github.com/react-native-webrtc/react-native-webrtc/issues/962